tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/2d.type.exists.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.type.exists</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.type.exists</h1>
<p class="desc">The 2D context interface is a property of 'window'</p>

<p class="notes">Defined in "Web IDL" (draft)
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("The 2D context interface is a property of 'window'");
_addTest(function(canvas, ctx) {

_assert(window.CanvasRenderingContext2D, "window.CanvasRenderingContext2D");


});
</script>
